+++ repo/src/main/java/com/parashift/onlyoffice/Prepare.java
@@ -158,6 +158,9 @@ public class Prepare extends AbstractWebScript {
                 responseJson.put("mime", mimetypeService.getMimetype(docExt));
                 responseJson.put("demo", configManager.demoActive());
                 responseJson.put("documentType", documentType);
+                responseJson.put("share", util.getShareUrl());
+                responseJson.put("saveas", util.getSaveAsUrl());
+                responseJson.put("currentPath", util.getCurrentPath(nodeRef));
 
                 logger.debug("Sending JSON prepare object");
                 logger.debug(responseJson.toString(3));