+++ share/src/main/resources/alfresco/site-webscripts/com/parashift/onlyoffice-edit.get.js
@@ -13,3 +13,6 @@ model.docTitle = pObj.config.document.title;
 model.config = JSON.stringify(pObj.config);
 model.demo = pObj.demo;
 model.documentType=pObj.documentType;
+model.share = pObj.share;
+model.saveas = pObj.saveas;
+model.currentPath = pObj.currentPath;