+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 #
 ## Added
+- save copy as... from document editor
 - creating new file from document editor
 - change favicon in editor by document type
 - detecting mobile browser